Made notes about how the `headers` attribute is passed to the Pypdl object and where it gets used.
It does seem like the `self._kwargs` dictionary is passed to `Multidown`, which then modifies the `headers` attribute. This is probably where the header gets modified.

When a new download begins the self._reset() function is called, which interestingly does *not* reset self._kwargs. Performing a reset of self._kwargs is difficult though, since the original parameters are not saved anywhere.

I've noticed that self._get_header() has very little exception handing and so have put a logger around it. Most notably the function will return `None` if both of the requests return non-200 status codes that are *also less than 400* (see `aiohttp` Docs here: https://docs.aiohttp.org/en/stable/client_reference.html#aiohttp.ClientResponse.raise_for_status).

This brings credit to the claim that there is an issue here, as the MDN docs (https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/206) indicate that status code `206` will be returned if 'range' is supplied. If 'range' is indeed sticky on a second request, *and* the server returns a code like `206` (instead of ignoring the header and returning `200` like I would expect some servers would), then this code definitely falls through and returns nothing, which would then propagate to a None reference somewhere down the line, leading to an unhandled exception. Ouch.

Only problem now is finding a server in the wild that does this. They were using nginx, but it's not like every server on the internet is just going to tell the truth in their 'server' response header...