Got it! Bug occurred as hypothesised. The first image in the download list (3eb0ebb8b3a93515fa070f6be303527c48ffeed1.jpg) saved exactly as expected, but when Pypdl attempted to pull the second file, it passed the 'range' header in the request, and so the server responded with `206` instead of `200`. Check the included log file for details.
I tried a lot of different social media sites first but all of them either required additional cookies or tokens (that might expire) and / or they responded with `200` instead of `206`. Needed a server that directly returns large files without authentication and figured something booru-powered might do the trick.

Not going to include the downloaded image for copyright reasons.