So, there are a couple of flaws in how `self._kwargs` is utilised and passed around, and I believe there are a few modifications to be made in certain places, both to fix the current bug *and* prevent a similar bug from happening in the future.
The main flaw is that self._kwargs is passed down by reference instead of value (since it is a dictionary and this is Python, of course). A callee may modify `self._kwargs` during its procedure (such as in the case of `Multidown()` adding the `range` header) and this is ultimately what causes issues when creating the next request.

This can be fixed by:
 - Not trusting the callee to properly handle `self._kwargs` and instead pass a `deepcopy()` to the callee. This is slow, but ensures that the dictionary passed to `Multidown()` can be modified at will. As demonstrated in this commit, this does fix the bug (and again proves its existence) but may not be suitable for a downloader whose main purpose is to be performant.
 - While this does fix the immediate bug, it does not prevent such another bug from occurring in a later extension of the program.