So, there are a couple of flaws in how `self._kwargs` is utilised and passed around, and I believe there are a few modifications to be made in certain places, both to fix the current bug *and* prevent a similar bug from happening in the future.
The main flaw is that self._kwargs is passed down by reference instead of value (since it is a dictionary and this is Python, of course). A callee may modify `self._kwargs` during its procedure (such as in the case of `Multidown()` adding the `range` header) and this is ultimately what causes issues when creating the next request.

This can be fixed by:
 - Removing the `kwargs['headers']['range']` attribute when inside `Multidown()`.
   - While faster than performing a `deepcopy()`, this still does not fix the design flaw, and if `Multidown().download()` fails, then `Multidown.worker()` won't be able to clean up 'range' header!
   - This method of fix by popping 'range' highlights the core problem for this type of fix, in that if execution aborts then the fix cannot be applied. This flaw exists regardless of the position that this patch is inserted at.