So, there are a couple of flaws in how `self._kwargs` is utilised and passed around, and I believe there are a few modifications to be made in certain places, both to fix the current bug *and* prevent a similar bug from happening in the future.
The main flaw is that self._kwargs is passed down by reference instead of value (since it is a dictionary and this is Python, of course). A callee may modify `self._kwargs` during its procedure (such as in the case of `Multidown()` adding the `range` header) and this is ultimately what causes issues when creating the next request.

This can be fixed by:
 - Explicitly resetting `self._kwargs` when `Pypdl._reset()` is called.
   - While this still requires a `deepcopy()` operation every time the download is reset, it does also guard against any future mishandling or callee modification of `self._kwargs`!
 - I believe this has the most coverage of any fix, but the performance hit should be noted. I don't exactly have the internet speeds to notice a difference, but others might.