So, there are a couple of flaws in how `self._kwargs` is utilised and passed around, and I believe there are a few modifications to be made in certain places, both to fix the current bug *and* prevent a similar bug from happening in the future.
The main flaw is that self._kwargs is passed down by reference instead of value (since it is a dictionary and this is Python, of course). A callee may modify `self._kwargs` during its procedure (such as in the case of `Multidown()` adding the `range` header) and this is ultimately what causes issues when creating the next request.

This can be fixed by:
 - Disconnecting the reference between the original `kwargs['headers']` by creating a shallow copy of the entry.
   - While this doesn't address the core issue, this is probably the fastest and most stable fix that can be added inside `Multidown()`. The worker function can add new headers at will without affecting the original headers supplied by the caller, and the performance impact should be negligible due to the use of a shallow copy. Those who do not provide custom headers should also experience negligible performance impact, as there is only 1 additional O(1) conditional per worker.